Resolve validation warnings on Quoted references
Warning: Quoted references are deprecated

  on external-dns.tf line 40, in resource "helm_release" "external-dns":
  40:     depends_on = ["kubernetes_cluster_role_binding.tiller"]

In this context, references are expected literally rather than in quotes.
Terraform 0.11 and earlier required quotes, but quoted references are now
deprecated and will be removed in a future version of Terraform. Remove the
quotes surrounding this reference to silence this warning.

(and 3 more similar warnings elsewhere)