EKS: move k8s_version from var to local
(1) In preparation for having automated selection of k8s version, remove
this from user options.
(2) QA noted that the move to 1.16 was in fact detrimental, and we
should stick with 1.15. I've observed intermittent failures on 1.16 and
continue to invesigate.
(3) Consistent with design of GKE, where k8s version is not exposed.